+++ meinberlin/react/projects/ProjectsList.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import django from 'django'
 import ProjectTile from './ProjectTile'
 import { classNames } from 'adhocracy4'
-
+import Spinner from '../contrib/Spinner'
 const translated = {
   showCompletedProjectsHeading: django.gettext('Show completed projects too?'),
   showCompletedProjectsBody: django.gettext('Expand your search to view completed projects and discover more possibilities.'),
@@ -10,12 +10,28 @@ const translated = {
 }
 
 const ProjectsList = ({
-  projects,
+  projects: initialProjects,
+  projectsUrl,
   topicChoices,
   isHorizontal,
   searchCompletedProjects,
   showSearchCompletedProjectsButton
 }) => {
+  const [projects, setProjects] = React.useState(initialProjects || [])
+  const [loading, setLoading] = React.useState(!initialProjects)
+
+  React.useEffect(() => {
+    if (projectsUrl) {
+      fetch(projectsUrl)
+        .then(response => response.json())
+        .then(data => {
+          setProjects(data)
+          setLoading(false)
+        })
+    }
+  }, [projectsUrl])
+
+  if (loading) return <Spinner />
   return (
     <ul className={classNames('projects-list', isHorizontal ? 'projects-list--horizontal' : 'projects-list--vertical')}>
       {